Refactor to avoid using UserStore for authentication
Because the UserStore implementation is configurable and it is designed
to be used with authorization it should be possible to have a UserStore
which does not provide authentication.
For the default implementation of login module for Jaas authentication
a separate file based user registry is used instead of the configurable
UserStore.